refactor Refinable class
